chore(jni): re-registry native methods
