+++ app/src/main/jni/librime_jni/key_table.cc
@@ -1,14 +1,18 @@
 #include "key_table.h"
 #include <rime/key_table.h>
 
-jint get_modifier_by_name(JNIEnv *env, jobject thiz, jstring name) {
+extern "C"
+JNIEXPORT jint JNICALL
+Java_com_osfans_trime_core_Rime_get_1modifier_1by_1name(JNIEnv *env, jclass thiz, jstring name) {
   const char* s = name == NULL ? NULL : env->GetStringUTFChars(name, NULL);
   int keycode = RimeGetModifierByName(s);
   env->ReleaseStringUTFChars(name, s);
   return keycode;
 }
 
-jint get_keycode_by_name(JNIEnv *env, jobject thiz, jstring name) {
+extern "C"
+JNIEXPORT jint JNICALL
+Java_com_osfans_trime_core_Rime_get_1keycode_1by_1name(JNIEnv *env, jclass thiz, jstring name) {
   const char* s = name == NULL ? NULL : env->GetStringUTFChars(name, NULL);
   int keycode = RimeGetKeycodeByName(s);
   env->ReleaseStringUTFChars(name, s);
